Schedule Import: Accept checkbox gating, MatchNote chip for alignment audit, campus-aware matcher with TC_Employees post-align, bigint SchoolId fix

+++ src/NewHeights.TimeClock.Data/Entities/Staff.cs
@@ -20,8 +20,11 @@ public class Staff
     /// (Jose Lagunas, for example) have a separate Staff row per campus, and
     /// the importer needs to prefer the row whose SchoolId matches the campus
     /// being imported. Joins to Campus.PowerSchoolId.
+    ///
+    /// Typed as long? because the Staff.SchoolId column is bigint. Comparisons
+    /// against Campus.PowerSchoolId (int?) widen automatically at the call site.
     /// </summary>
-    public int? SchoolId { get; set; }
+    public long? SchoolId { get; set; }
 
     public string FullName => $"{FirstName} {LastName}".Trim();
 }
